#20 feat(front) ajusta componentizacao da homePage e cria componente dos graficos
