+++ front/src/App.tsx
@@ -5,7 +5,7 @@ import { LoginPage } from './pages/LoginPage';
 import { NotificationsProvider } from '@mantine/notifications';
 import { Navigation } from './components/Navigation';
 import { ModalProvider } from './context/ModalContext/ModalContext';
-import { TransacaoPage } from './pages/TransacaoPage';
+import { Home } from './pages/Home';
 import { CustomPage } from './pages/CustomPage';
 
 function App() {
@@ -19,7 +19,7 @@ function App() {
         >
           <BrowserRouter>
             <Routes>
-              <Route path="/" element={<TransacaoPage />} />
+              <Route path="/" element={<Home />} />
               <Route path="/custom" element={<CustomPage />} />
               <Route path="/login" element={<LoginPage />} />
               <Route path="/cadastro" element={<CadastroPage />} />
