+++ front/src/components/MainDashboard/components/InfoCards/components/Card/index.tsx
