+++ front/src/components/MainDashboard/components/InfoCards/components/Card/styles.ts
