+++ front/src/components/MainDashboard/components/TransacaoList/TransacaoItem/styles.ts
