--- front/src/components/TransacaoList/index.tsx
@@ -1,323 +1,0 @@
-import {
-  Alert,
-  Box,
-  Button,
-  Center,
-  Divider,
-  Group,
-  Loader,
-  Paper,
-  ScrollArea,
-  TextInput,
-  ThemeIcon,
-  Title,
-} from '@mantine/core';
-import { useDisclosure } from '@mantine/hooks';
-import { useContext, useEffect, useState } from 'react';
-import {
-  AiOutlineInfoCircle,
-  AiOutlinePlus,
-  AiOutlineSearch,
-} from 'react-icons/ai';
-import { BiChart } from 'react-icons/bi';
-import { BsGraphUp } from 'react-icons/bs';
-import { GiMoneyStack } from 'react-icons/gi';
-import { useQuery } from 'react-query';
-import AuthContext from '../../context/AuthContext/AuthContext';
-import { useModalController } from '../../context/ModalContext/ModalContext';
-import { useMonthController } from '../../context/MonthContext/MonthContext';
-import { findAllCategorias } from '../../services/categoria';
-import { findAllTransacao, TransacaoResponse } from '../../services/transacao';
-import { InfoCards } from '../InfoCards';
-import { SeletorMes } from './components/SeletorMes';
-import { TransacaoItem } from './components/TransacaoItem';
-import { TipoTransacaoEnum } from './components/TransacaoModals/constants';
-import { CreateTransacaoModal } from './components/TransacaoModals/CreateTransacaoModal';
-import { EditTransacaoModal } from './components/TransacaoModals/EditTransacaoModal';
-import CustomChart from './CustomChart';
-import { getTransacaoByTipo, getValues } from './formatter';
-
-export function TransacaoList() {
-  const [openedCreate, handlersCreate] = useDisclosure(false);
-  const [openedEdit, handlersEdit] = useDisclosure(false);
-
-  const { onSetId } = useModalController();
-  const { date } = useMonthController();
-  const { token } = useContext(AuthContext);
-
-  const { data, isLoading, error } = useQuery(['transacoes'], () => {
-    return findAllTransacao(token.token);
-  });
-
-  const { data: dateCategorias } = useQuery(['categorias'], () => {
-    return findAllCategorias(token.token);
-  });
-
-  const [receitas, setReceitas] = useState<TransacaoResponse[]>([]);
-  const [despesas, setDespesas] = useState<TransacaoResponse[]>([]);
-
-  const handleOpenEditModal = (id: string) => {
-    onSetId(id);
-    handlersEdit.open();
-  };
-
-  useEffect(() => {
-    if (data) {
-      setReceitas(
-        getTransacaoByTipo(TipoTransacaoEnum.RECEITA, data.data, date)
-      );
-      setDespesas(
-        getTransacaoByTipo(TipoTransacaoEnum.DESPESA, data.data, date)
-      );
-    } else {
-      setReceitas([]);
-      setDespesas([]);
-    }
-  }, [data, date]);
-
-  return (
-    <Box
-      style={{
-        minWidth: '100%',
-        padding: '1rem',
-      }}
-    >
-      <Group
-        style={{
-          justifyContent: 'space-between',
-          marginBottom: '2rem',
-        }}
-      >
-        <Button
-          color="blue"
-          leftIcon={<AiOutlinePlus />}
-          onClick={() => handlersCreate.open()}
-        >
-          Adicionar nova transação
-        </Button>
-        <SeletorMes />
-        <TextInput
-          icon={<AiOutlineSearch size={18} />}
-          placeholder="Pesquisar"
-          disabled
-        />
-      </Group>
-
-      <Divider mt="2rem" size="md" color="blue" />
-
-      <Group align="center" mt="2rem">
-        <Title order={2} align="center">
-          Resumo
-        </Title>
-        <ThemeIcon size={30}>
-          <BsGraphUp size={30} />
-        </ThemeIcon>
-      </Group>
-      <InfoCards
-        isLoading={isLoading}
-        values={
-          data && data.data.length > 0 ? getValues(data.data, date) : null
-        }
-      />
-
-      <Group align="center" mt="2rem">
-        <Title order={2} align="center">
-          Transações
-        </Title>
-        <ThemeIcon size={30}>
-          <GiMoneyStack size={30} />
-        </ThemeIcon>
-      </Group>
-      <Box
-        style={{
-          minWidth: '100%',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          marginTop: '2rem',
-          gap: '2rem',
-        }}
-      >
-        <Paper
-          style={{ backgroundColor: '#D1f7d6', flex: 1 }}
-          shadow="md"
-          p="1rem"
-        >
-          <ScrollArea style={{ height: 500 }}>
-            <>
-              {isLoading && (
-                <Box
-                  style={{
-                    height: '500px',
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                  }}
-                >
-                  <Loader />
-                </Box>
-              )}
-              {!isLoading && error && (
-                <Box
-                  style={{
-                    height: '500px',
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                  }}
-                >
-                  <Alert
-                    icon={<AiOutlineInfoCircle size={20} />}
-                    title="Ops!"
-                    color="red"
-                  >
-                    Aparentemente alguma coisa deu errado, tente novamente
-                  </Alert>
-                </Box>
-              )}
-
-              {receitas && receitas.length > 0 && (
-                <Group>
-                  {receitas.map((transacao) => (
-                    <TransacaoItem
-                      key={transacao.id}
-                      data={transacao}
-                      onOpenEdit={handleOpenEditModal}
-                    />
-                  ))}
-                </Group>
-              )}
-              {!isLoading && !error && receitas && receitas.length === 0 && (
-                <Box
-                  style={{
-                    height: '500px',
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                  }}
-                >
-                  <Alert
-                    icon={<AiOutlineInfoCircle size={20} />}
-                    title="Ops!"
-                    color="blue"
-                  >
-                    Nenhuma informação cadastradada
-                  </Alert>
-                </Box>
-              )}
-            </>
-          </ScrollArea>
-        </Paper>
-
-        <Paper
-          style={{ backgroundColor: '#Ffd3d3', flex: 1 }}
-          shadow="md"
-          p="1rem"
-        >
-          <ScrollArea style={{ height: 500 }}>
-            <>
-              {isLoading && (
-                <Box
-                  style={{
-                    height: '500px',
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                  }}
-                >
-                  <Loader />
-                </Box>
-              )}
-              {!isLoading && error && (
-                <Box
-                  style={{
-                    height: '500px',
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                  }}
-                >
-                  <Alert
-                    icon={<AiOutlinePlus size={20} />}
-                    title="Ops!"
-                    color="red"
-                  >
-                    Aparentemente alguma coisa deu errado, tente novamente
-                  </Alert>
-                </Box>
-              )}
-
-              {data && data.data.length > 0 && (
-                <Group>
-                  {despesas.map((transacao) => (
-                    <TransacaoItem
-                      key={transacao.id}
-                      data={transacao}
-                      onOpenEdit={handleOpenEditModal}
-                    />
-                  ))}
-                </Group>
-              )}
-              {!isLoading && !error && despesas && despesas.length === 0 && (
-                <Box
-                  style={{
-                    height: '500px',
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                  }}
-                >
-                  <Alert
-                    icon={<AiOutlinePlus size={20} />}
-                    title="Ops!"
-                    color="blue"
-                  >
-                    Nenhuma informação cadastradada
-                  </Alert>
-                </Box>
-              )}
-            </>
-          </ScrollArea>
-        </Paper>
-      </Box>
-
-      <Group align="center" mt="2rem">
-        <Title order={2} align="center">
-          Dashboards
-        </Title>
-        <ThemeIcon size={30}>
-          <BiChart size={30} />
-        </ThemeIcon>
-      </Group>
-
-      <Center>
-        <Title>Receitas</Title>
-        <CustomChart />
-      </Center>
-
-      <Center>
-        <CustomChart />
-        <Title>Despesas</Title>
-      </Center>
-
-      {/* Modals */}
-      {dateCategorias?.data && (
-        <>
-          <CreateTransacaoModal
-            categorias={dateCategorias.data}
-            isOpen={openedCreate}
-            onClose={handlersCreate.close}
-          />
-          {data && data.data.length > 0 && (
-            <EditTransacaoModal
-              categorias={dateCategorias.data}
-              isOpen={openedEdit}
-              onClose={handlersEdit.close}
-              transacaoList={data.data}
-            />
-          )}
-        </>
-      )}
-    </Box>
-  );
-}
