+++ front/src/pages/Home/index.tsx
@@ -1,11 +1,11 @@
 import { useContext, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Navigation } from '../../components/Navigation';
-import { TransacaoList } from '../../components/TransacaoList';
+import { MainDashboard } from '../../components/MainDashboard';
 import AuthContext from '../../context/AuthContext/AuthContext';
 import { MonthProvider } from '../../context/MonthContext/MonthContext';
 
-export function TransacaoPage() {
+export function Home() {
   const navigate = useNavigate();
   const { token } = useContext(AuthContext);
 
@@ -18,7 +18,7 @@ export function TransacaoPage() {
   return (
     <Navigation>
       <MonthProvider>
-        <TransacaoList />
+        <MainDashboard />
       </MonthProvider>
     </Navigation>
   );
